move tasks badge to its own icon
