+++ components/common/PageLayout/components/Header/components/NotificationsBadge/index.tsx
@@ -1,0 +1,3 @@
+import NotificationsBadge from './NotificationsBadge';
+
+export default NotificationsBadge;
