+++ components/common/PageLayout/components/Header/components/Snapshot/InputVotingStrategies.tsx
