+++ tsconfig.json
@@ -36,6 +36,7 @@
   ],
   "include": [
     "**/*.ts",
-    "**/*.tsx"
+    "**/*.tsx",
+    "components/common/PageLayout/components/Header/components/NotificationsBadge"
   ]
 }
